requirements.txt
selenium==4.15.0
pytest==8.4.2
allure-pytest==2.15.0